Re-arranging the helper subroutines.
